Move EventItem to own component
